feat: remove `feeCurrency`, `gatewayFee`, and `gatewayFeeRecipient` properties from Celo transaction receipts (#1500)
* test(formatter): adds failing tests.
The transaction receipt returned by an RPC node never returns a feeCurrency field.
It's always `undefined`.

* fix(celo): removes `feeCurrency` field from the transaction receipt formatter, and adjusts tests

* test(celo): updates transaction receipt snapshots

* chore(changeset): add changeset

* feat(celo): removes custom transaction receipt formatter

Including now redundant tests, and types.

* test(custom chain): replacing celo with zkSync in test

Celo doesn't have a custom transaction receipt formatter, the existing test passes trivially. We replaced it with zkSync, which has a custom transaction receipt formatter, to maintain test coverage.

* chore(celo): updates snapshots

* chore: linting

* chore(celo): undo getTransaction test to check code coverage

Previously we replaced celo with zkSync here in an attempt to keep code coverage up. We assumed that the test was unnecessary, but maybe it was necessary after all.

* Revert "chore: linting"

This reverts commit 93479b0f1366c4cdd42512c0484761d96e20b980.

* Revert "chore(celo): updates snapshots"

This reverts commit f1f7690da0bcc8d18fb65234147d35d0a12ee285.